- Changes which enable each reaction to have a list of species indices and access to the species list (instead of just their names) so that reaction parsing, validation, and type determination can be completed by each reaction

+++ src/kinetics/StoichiometryManager.cpp
@@ -9,7 +9,7 @@ using namespace Mutation::Numerics;
 namespace Mutation {
     namespace Kinetics {
 
-void StoichiometryManager::addReaction(const int rxn, const vector<size_t> &sps)
+void StoichiometryManager::addReaction(const int rxn, const vector<int> &sps)
 {
     switch (sps.size()) {
         case 1:
